requirements.txt
beautifulsoup4==4.11.2
Requests==2.31.0
Pysocks==1.7.1
